fix: Prevent FFmpeg from consuming stdin
By default, FFmpeg runs in an "interactive" mode which lets the user
interact with FFmpeg while it is running by pressing keys.  When
FFmpeg is being run as a subprocess or background task, this feature
needs to be disabled to prevent FFmpeg from unintentionally capturing
stdin.  FFmpeg's default behavior caused an odd bug with tracklistify
when processing certain types of files (.mp4 files) which would
trigger this behavior and break the shell's echo.

Per the FFmpeg FAQ, this feature can be disabled by either using the
-nostdin flag, or by redirecting /dev/null (or NUL) to stdin.  I
chose the former since it is more portable.

https://ffmpeg.org/faq.html#How-do-I-run-ffmpeg-as-a-background-task_003f